# Conflicts:
#	TestData/DataAll.xls

+++ src/com/qingzi/api/meeting/admitMeeting.java
@@ -9,49 +9,43 @@ import com.qingzi.testUtil.RequestDataUtils;
 import com.qingzi.testUtil.StringUtils;
 import io.restassured.path.json.JsonPath;
 import io.restassured.response.Response;
+import net.sf.json.JSONObject;
 import org.bson.Document;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 
 /**
  * 
- * @ClassName:  auth   
- * @Description:TODO 获取解决方案  token（三方服务器转发客户端请求）
- * 
- *
+ * @ClassName:  admit
+ * @Description:等候同意后获取会议信息和媒体信息
  * @author: wff
- * @date:  2021年12月6日16:42:27
+ * @date:   2021年12月7日18:47:19
  * @Copyright:
  */
-public class getTokenByOther extends QZ implements API {
+public class admitMeeting extends QZ implements API {
 	
 	public String parameter; //参数集合
-	public String BUid; //	用户三方唯一标识
-	public  String userName; //邮箱
+	public String meetingId; //会议Id
 
 	@Override
 	public void initialize(HashMap<String, Object> data) {
-		parameter = MapUtil.getValue("parameter", data);
-		userName = MapUtil.getParameter(parameter,"userName").trim();
-		//根据邮箱查询BUid
-		Document docs =  MongoDBUtil.findByid(data, "crystal", "bucUser", "username", userName);
-		BU_id = docs.getObjectId("_id").toString();
-//		System.out.println(BU_id);
+
 	}
 
 	@Override
 	public HashMap<String, Object> handleInput(HashMap<String, Object> data) {
 		parameter = MapUtil.getValue("parameter", data);
-		
-		BUid = MapUtil.getParameter(parameter,"BUid").trim();
-		if(!BUid.equals("") && BUid.equals("code")){
-			BUid = BU_id;
-			parameter = parameter.replace("\"BUid\":code", "\"BUid\":\""+ BUid + "\"");
+
+
+
+		meetingId = MapUtil.getParameter(parameter,"meetingId").trim();
+		if(!meetingId.equals("") && meetingId.equals("code")){
+			meetingId = meeting_Id;
+			parameter = parameter.replace("\"meetingId\":code", "\"meetingId\":\""+ meetingId + "\"");
 		}
-
-		String[] parameter2 = parameter.split(",");
-
-		data.put("parameter", parameter2[0]);
+		
+		data.put("parameter", parameter);
 		return data;
 	}
 
@@ -59,12 +53,13 @@ public class getTokenByOther extends QZ implements API {
 	public Response SendRequest(HashMap<String, Object> data, String Url,
 			String Request) {
 		HashMap<String, String> headers = new HashMap<String, String>();
+		//需要调用奇瑞域名才能获取
+		headers.put("SUserToken",s_UserToken_Other.get("firstToken"));
 		headers.put("appId",appId);
-		headers.put("appSecret","");
-		headers.put("dev","phone");
+		headers.put("dev",dev);
 		
 		MyRequest myRequest = new MyRequest();
-		myRequest.setUrl("/moms/auth/v1/getToken");
+		myRequest.setUrl(Url);
 		myRequest.setHeaders(headers);
 		myRequest.setRequest(Request);
 		myRequest.setParameter(parameter);
@@ -128,24 +123,34 @@ public class getTokenByOther extends QZ implements API {
 //				if (!isProduct) {
 //					
 //				}
-				//目前访问域名直接到奔奔内部接口没有经过奇瑞外部包装（现在返回字段为id），后期会更换域名到时候接口返回字段为s_UserToken
-				s_UserToken_Other = new HashMap<>();
-				s_UserToken_Other.put("firstToken",jp.getString("data.s_UserToken"));
-				System.out.println("s_UserToken_Other = " + s_UserToken_Other.get("firstToken"));
-				userAccountIdByOther = jp.getString("data.accountId");
-				MR_Id = jp.getString("data.MRId");
-				
-				
-				/*if (data.get("CleanDB") != "" && data.get("CleanDB").equals("Y")) {
+				/*//接口返回meetingid
+				meeting_Id = jp.getString("data.meetingId");
+				m_Id = jp.getString("data.mId");
+				sdk_AccountId = jp.getString("data.sdkAccountId");
+				sdk_RoomId = jp.getString("data.sdkRoomId");*/
+				//参会人userAccountId
+				sdkAccountIdByOther = jp.getString("data.mediaInfo.sdkAccountId");
+				sdkRoomId = jp.getString("data.mediaInfo.sdkRoomId");
+
+				if (data.get("CleanDB") != "" && data.get("CleanDB").equals("Y")) {
+					
 					//先查询该用户创建的个人会议
-					Document doc =  MongoDBUtil.findByid(data, "crystal", "usrmgrAccount", "BUid", "feifei");
+					Document doc =  MongoDBUtil.findByid(data, "crystal", "usrmgrAccount", "BUid", BU_id);
 					String personalRoomId = doc.getString("personalRoomId");
 //					System.out.println(personalRoomId);
+					//删除企业
+					MongoDBUtil.deleteByid(data, "crystal", "usrmgrEnterprise", "name", enterprise_name);
 					//删除个人注册后创建的个人会议室
 					MongoDBUtil.deleteByid(data, "crystal", "mcmuMeetingRoom", "_id", personalRoomId);
 					//删除会前注册信息
-					MongoDBUtil.deleteByid(data,"crystal","usrmgrAccount","BUid","feifei");
-				}*/
+					MongoDBUtil.deleteByid(data,"crystal","usrmgrAccount","BUid", BU_id);
+					//删除会议记录
+					MongoDBUtil.deleteByid(data, "crystal", "mtmgrMeetingAuthLog", "meetingId", meetingId);
+					//删除参会表
+					MongoDBUtil.deleteByid(data, "crystal", "mtmgrMeetingParticipant", "accountId", userAccountId);
+					//删除新建会议
+					MongoDBUtil.deleteByid(data, "crystal", "mtmgrMetting", "title", title_meeting);
+				}
 			}
 			
 		}
